Temporal documentation update, part 2 (#4354)
<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

More updates to the temporal documentation started in #4353.

This PR contains the updates to PlainTime and PlainDateTime.

The major updates are made on the respective commits for the built-ins.
There were a couple follow up commits to add some fixes.