`Fetch` API (beta) (#4338)
Introduce an implementation of `fetch`
([mdn](https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API)) to
`boa_runtime`. It is optional and covers many basic use cases but not
all.

Other house keeping and yak shaving for this PR:

- Some fixes and improvements
- The new `fetch` is added to boa's CLI, using a blocking `reqwest`
client. We publish a blocking client implementation of `Fetcher` here to
allow for non-async applications to use it. A non-blocking client should
be added.
- Broaden the implementation of `TryIntoJsResult` to also include any
types that can be `TryIntoJs` already. Before it was limited to
`Into<JsValue>`.
- Add missing implementations of `TryIntoJs` for `JsPromise`, any `T:
Class + Clone` (since it's a trivial implementation), and
`isize`/`usize`.
- Update `JsPromise::from_future` to include an `AsyncFnOnce` that
receives a context.
- Add `Nullable::unwrap_or_default()`, which is used in this PR for
simplicity.
- Add an implementation of `Trace` and `Finalize` for `Either`
(conditional on existing feature)
- Fix the interface of `clearTimeout` to allow for nulls (now that
`Option<>` only supports `undefined`). This fixes a couple WPT.
- Clarify how to initialize `boa_runtime` now that it's becoming more
complex.
- Add support for `fetch` to WPT tests. Unfortunately, it's basically
failing everything and requires a webserver to run locally, so WPT will
not run on CI just yet and all fetch tests are ignored for now.
- Add support for generics in `boa_module`.
- Add support for a `boa_register()` function on `boa_modules` for those
who want to register the module's content on the context/realm.

---------

Co-authored-by: José Julián Espina <jedel0124@gmail.com>
Co-authored-by: Haled Odat <8566042+HalidOdat@users.noreply.github.com>