Add (optionally) Float16Array and f16round() support and add JsUint8ClampedArray (#4364)
For some reason the second one was missing from Rust side.

The reason to gate this behind a feature is that f16 is niche as best
when using Boa in an application. Adding dependencies based on barely
used features just makes the code bloated.

Enabling the test262 tests and the feature in boa_tester passes all the
relevant tests.