Fix clippy lints for Rust 1.89 (#4368)
<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

Fixes a variety of cargo clippy lints for the new version.

This also removes quite a few clones in `boa_engine`! We should probably
try to watch out for that pattern in the future and make use of
`std::slice::from_ref`