Implement `Atomics.waitAsync` (#4339)
Phew, this was a bit of an adventure.

This implements `Atomics.waitAsync`, which is the last missing
synchronization primitive we needed to implement to have 100%
conformance on that builtin. The implementation is not the most optimal
one; we could definitely use some channels to communicate which waiters
were notified instead of having to iterate through all of them, but IMO
it looks fairly decent for a first pass.