Fix broken PlainMonthDay.from tests (#4389)
<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

This Pull Request is related to ongoing work for #1804.

This updates `to_month_day` to better align with the specification and
fixes some bugs that was causing some broken tests.