Fix final Temporal.Instant test in built-ins test suite (#4388)
<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

This Pull Request closes #4074.

It is also related to ongoing work for #1804.

It fixes the final outstanding test for Temporal.Instant. We will still
be missing intl402 support for `toLocaleString`, but that is dependent
on the `Intl.DateTimeFormat` implementation.