Implement `Math.sumPrecise` (#4383)
<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

This PR implements
[Math.sumPrecise](https://tc39.es/ecma262/#sec-math.sumprecise)
functionality.

This uses the `xsum` crate for the implementation, so that crate has
been added behind a feature flag for the dependency and the dependency
has been added to `boa_engine`'s defaults.