Remove alignment of property keys with identation (#4390)
This fixes some visual issues like alignment of brackets, and also make
it more consistent when keys are long (it does not take much, only need
to be longer than ident which is 4).

Since most of our tests have single letter keys this never showed up as
much of a problem, but in the real world it becomes annoying.

Before:

```
{
hello: "world",
   z: {
       a: 1
    }
}
```

After:

```
{
    hello: "world",
    z: {
        a: 1
    }
}
```