Update `Math.sumPrecise` to remove `SummationState` (#4392)
<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

Follow up to #4383.

As noted, there may be further optimizations we could do in the future
with Array's where the length is known and `XsumSmall` and `XsumLarge`
are used directly vs. `XsumAuto`.