Fix index access in `JumpTable` instruction (#4372)
Prior to this PR we accessed `JumpTable` index by popping an element of
the stack, with the introduction of #3798 the mechanisms for clearing
the stack have been removed and we cannot ensure that the index is the
last element. This PR switches the JumpTable instruction to use a
register to access the index.