Unblock Context::run_jobs when no timeouts need to be run
Prior to this PR, `run_jobs()` would block as long as there would
be timeout jobs, even if they were in the future. A good example
is using `setInterval` would run an infinite loop in `run_jobs()`.

This PR unblock `run_jobs()` when the jobs do not need to be run
right now.

The new tests would infinite loop prior to this PR.