Unblock Context::run_jobs when no timeouts need to be run (#4416)
Prior to this PR, `run_jobs()` would block as long as there would be
timeout jobs, even if they were in the future. A good example is using
`setInterval` would run an infinite loop in `run_jobs()`.

This PR unblock `run_jobs()` when the jobs do not need to be run right
now.

The new tests would infinite loop prior to this PR.

This PR also moved the code for the CLI prompt to a separate thread so
the JavaScript code isn't waiting on prompts before executing tasks.
This means that intervals/timeouts/promises/etc return immediately and
will output their logs and move the prompt without interupting the user.