Fix the CLI when it is not attached to a TTY (#4424)
This a simple fix when the CLI is not attached to a TTY. In those cases
it would print an error `readline thread failed: ENOTTY: Not a
typewriter`.

This fixes it by doing three things;
1. by default, if no printers has been set, use stdout normally. This
makes sense as if any logging occurs before readline can be initialized,
we still want to see it.
2. move back starting the readline loop after args files/expressions are
handled.
3. if we cannot build a writer for some reason, don't stop the thread.
`readline` still works, but printers won't, so we just show logs on
stdout. This can happen if STDIN is a pipe.