Update to temporal_rs to 0.1 release (#4433)
<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

This is related to ongoing work for #1804

This PR updates Boa to the new 0.1 release of temporal_rs and
timezone_provider.

Notably, I've switched us over to `CompiledTzdbProvider` for now, which
will come with an increase binary size (I think the jiff_tzdb .dat file
is ~200 Kb). I am hoping to update this in the near future to be more
configurable though.