Allow printing test262 comment on PRs from forks (#4428)
Fixes our test262 action so that it can still print the test results
from PRs opened from forks.

One caveat is that any changes made to that action won't be reflected on
the PR that introduces the changes, because `pull_request_target` makes
it so that Github pulls code from `main` to run the action, instead of
pulling the code from the PR branch.