Fix unconditional `requires_function_scope` in constructor (#4470)
<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

This Pull Request fixes/closes #4400.

It changes the following:

- In `ScopeIndexVisitor`, the `requires_function_scope` flag is always
set to true for class constructors.

The root cause is that `compile_class` unconditionally flags the
constructor with `HAS_FUNCTION_SCOPE`. However, when all bindings are
local, `ScopeIndexVisitor` optimizes away the function scope. This
creates a discrepancy between the compile-time state and the runtime
state.

I'm not sure if this is the suitable approach. I attempted to remove the
`function_scope` during constructor handling in `compile_class`, but
this breaks other parts of the code. `CheckReturn` and
`SuperCallDerived` rely on reading `this` from the `environment`. So, it
seems we have to push the `function_scope` there.