Fix typos across the tree (#4512)
This has been done with the help of the wonderful
[typos](https://github.com/crate-ci/typos) tool, and manual review of
each individual change. A few changes are my own as well.

<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

This Pull Request fixes/closes no issue.

It changes the following:
- Many typos have been fixed to be proper English.
- Of course not all of them.