Allow TimeZoneProvider to be configurable via the ContextBuilder (#4447)
<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

This is related to work on #1804

It changes the following:

- Implements a `DynamicTimeZoneProvider`.
- Adds a method for setting a custom `TimeZoneProvider` on
`ContextBuilder`.

This is related to the general API, so please feel free to let me know
if you think this is the wrong approach.

Assuming we're happy with the result of this PR and its merged, I would
be fine removing Temporal from the `experimental` feature flags for the
next release. We still need to figure out a good way to make reading the
host system time zone accessible, but that's still an odd aspect of the
specification itself.