Improve strictness around `Map` and `Set` objects (#4497)
Fixes #4493.
Depends on #4495.

This PR improves strictness on the implementations of `Map` and `Set` by
ensuring borrows are as small as possible, and also by reimplementing
our lock structs to be a bit less sporadic about when they lock and
unlock the underlying map.

Additionally, it cleans up our implementations to use newer APIs such as
the `js_error` macro, `upcast/downcast`, etc.