fix: make Scope's unique_id truly unique and allow access to it
Right now, `Scope::unique_id` is always the same as the parent Scope's
`index + 1`. This means that if two different scopes are created as
children of the same scope, they will have the same `unique_id`, belying
its name.

In this PR, add `unique_id_ctr` field to the root scope, which controls
the identifiers that are given out to scopes in the tree. For simplicity
the counter field is in every `Scope`, but only the root scope's counter
field will be used to hand out unique identifiers.

Also make the field accessible via a getter function, as well as on
`BindingLocator`, so that `BindingLocator` and `Scope` can be
correlated.

Relates to #4519.