fix: make Scope's unique_id truly unique and add a getter for it (#4527)
Right now, `Scope::unique_id` is always the same as the parent Scope's
`index + 1`. This means that if two different scopes are created as
children of the same scope, they will have the same `unique_id`, belying
its name.

In this PR, add a new `ScopeContext` struct that all Scopes in an AST
share; this new struct contains a container which controls the
identifiers that are given out to scopes in the tree.

Also make the field accessible via a getter function, as well as on
`BindingLocator`, so that `BindingLocator` and `Scope` can be
correlated.

Relates to #4519.