Extract engine errors as their own `JsError` variant (#4534)
Related to #4515.

Extracts a new `EngineError` type that is a new variant of `JsError`
that cannot be converted to a `JsValue` using `into_erased`.

This also makes `into_erased` fallible for `JsError`, which IMO is fine
since we really want to make sure users handle engine errors gracefully.

EDIT: Unfortunately, this doesn't fully address #4515 because the stack
overflows on the CI profile. Will open an issue related to that.