Remove `Expression::FormalParameterList` (#4537)
<!---
Thank you for contributing to Boa! Please fill out the template below,
and remove or add any
information as you feel necessary.
--->

This Pull Request fixes #4403 .

It changes the following:

- Remove `Expression::FormalParameterList`
- Introduce a new structure, `FormalParameterListOrExpression`, as the
output for certain parsers (e.g., `MemberExpression`,
`UnaryExpression`).
- Throw an error early in cases where it's known that `CPEAAPL` cannot
be interpreted as a `FormalParameterList`.

I'm uncertain about this change. It adds complexity to the parser.