Change the internal structure of JsString for performance
This is the first in a series of PR to improve the performance of strings in Boa.

The first step was to introduce a new type of strings, `SliceString`, which contains a strong pointer to another string and start/end indices. This allows for very fast slicing of strings. This initially came at a performance cost by having an enumeration of kinds of strings. An intermediate experiment was introduced to have the kind be a tag on the internal JsString pointer. This still came as a cost as it required bit operations to figure out which function to call.

Finally, I moved to using a `vtable`. This helped with many points:
1. as fast as before. Before this PR, there was still a deref of a pointer when accessing internal fields.
2. we can now introduce many other types (which will come in their separate PRs).
3. this makes the code to clone/drop/as_str (and even construction) more streamline as each function is their own implementation.