Change the internal structure of JsString for performance (#4455)
This is the first in a series of PR to improve the performance of
strings in Boa.

The first step was to introduce a new type of strings, `SliceString`,
which contains a strong pointer to another string and start/end indices.
This allows for very fast slicing of strings. This initially came at a
performance cost by having an enumeration of kinds of strings. An
intermediate experiment was introduced to have the kind be a tag on the
internal JsString pointer. This still came as a cost as it required bit
operations to figure out which function to call.

Finally, I moved to using a `vtable`. This helped with many points:
1. as fast as before. Before this PR, there was still a deref of a
pointer when accessing internal fields.
2. we can now introduce many other types (which will come in their
separate PRs).
3. this makes the code to clone/drop/as_str (and even construction) more
streamline as each function is their own implementation.

The biggest drawback is now we are operating with a bit more pointer and
unsafe code. I made sure that all code was passing both debug, release
and miri tests on boa_string and boa_engine's builtins::string suite.

This `vtable` bring slightly better performance (depending on the test,
about 90-110%, but overall 2-3% better than the `main` branch), even
though we only directly improved string slicing operations.

The next suite of PRs are going to cover more string types:
1. since strings are constant and unmutable, include the length in the
vtable directly instead of relying on a function call.
2. also because of the above, splitting latin1 and utf-16 string kinds
to remove conditionals on a lot of sub functions.
3. have a concat string that makes concatenation instant, similar to how
slice improve concatenation

Also, there might still be some slicing that creates new strings from
existing ones without using SliceString. Those can be improved greatly.

Please note the `benches/scripts/strings/slice.js` test is twice as fast
with this branch than with `main`.

The performance table for this PR:

| Test | Main | PR  |
|---|-----|-----|
| Richards | 218 | 233 |
| DeltaBlue | 234 | 244 |
| Crypto | 184 | 208 |
| RayTrace | 502 | 485 |
| EarleyBoyer | 595 | 584 |
| RegExp | 88.1 | 83  |
| Splay | 905 | 837 |
| NavierStokes | 413 | 467 |
|---| --- | --- |
| TOTAL | 313 | 320 |